Added mechanism to share constants between C and assembler files
Use 'define()' macro inside of *.def.h files to define constants and
add the name of the %.def.h file to DEFH in /src/Makefile.  The build
system will generate a %.gen.h and a %.gen.s file which are able to
include into C files (%.gen.h) or assembler files (%.gen.s).